\name{vect_desc}
\alias{vect_desc}
\title{Plot and analyse a vector}
\usage{
vect_desc(vect , color)
}
\arguments{
\item{vect}{A vector that we want to plot and analyse}

\item{color}{The color of the plot}
}

\description{
This function produces the  plot and summary of a vector that defined as input
}


\examples{
a=c(2,10,8,2,10,2,1,15,12,9)

vect_desc(vect = a , color = "red" )
}


\author{
Kouki Mohamed
}
